Update resolvers for fixed schema

+++ graphql/resolvers/order/mutations.ts
@@ -5,14 +5,14 @@ const orderMutations = {
   createOrder: async (parent, { input }, { user }) => {
     try {
       const orderData = Object.assign({}, input, { user: user._id });
+
       const newOrder = new Order(orderData);
       const orderResult = await newOrder.save();
 
-      const orderCreator = await User.findById(user.id);
-      if (orderCreator) {
-        orderCreator.userOrders.unshift({ order: orderResult._id});
-        await orderCreator.save();
-      }
+      const orderCreator = await User.findOneAndUpdate(
+        { _id: user._id }, { $addToSet: { userOrders: orderResult._id } }
+      );
+      await orderCreator.save();
 
       return {
         message: 'Order created!',
@@ -24,13 +24,12 @@ const orderMutations = {
   },
   updateOrder: async (parent, { id, input }, context) => {
     try {
-      const { orderItems, orderTotal, deliveryCost, totalCost, deliveryAddress, deliveryDate, paymentResult, isPaid, paidAt } = input
+      const { product, deliveryCost, totalCost, deliveryAddress, deliveryDate, paymentResult, isPaid, paidAt } = input
 
       const order = await Order.findById({ _id: id });
 
       if (order) {
-        order.orderItems = orderItems;
-        order.orderTotal = orderTotal;
+        order.product = product;
         order.deliveryCost = deliveryCost;
         order.totalCost = totalCost;
         order.deliveryAddress = deliveryAddress;
@@ -52,17 +51,15 @@ const orderMutations = {
       return error
     }
   },
-  deleteOrder: async (parent, { id }, context) => {
+  deleteOrder: async (parent, { id }, { user }) => {
     try {
       const order = await Order.findById({ _id: id });
 
-      if (order) {
-        const orderCreator = await User.findById(order.user);
-        if (orderCreator) {
-          const orderIndex = orderCreator.userOrders.indexOf({ order: order._id });
-          orderCreator.userOrders.splice(orderIndex, 1);
-          await orderCreator.save();
-        }
+      if (order && user.role === 'ADMIN') {
+        const orderCreator = await User.findOneAndUpdate(
+          { _id: order.user }, { $pull: { userOrders: order._id } }
+        );
+        await orderCreator.save();
 
         const deletedOrder = await Order.findOneAndDelete({ _id: id });
